+++ src/pages/OnBoarding/OnBoardingStep.styled.jsx
