+++ src/pages/OnBoarding/Step/LandingStep.jsx
@@ -1,9 +1,9 @@
 import React, { useEffect, useState } from 'react'
-import * as Styled from './OnBoardingStep.styled'
+import * as Styled from '../OnBoardingStep.styled'
 import { ReactMarkdown } from 'react-markdown/lib/react-markdown'
 import YnputConnector from '../../SettingsPage/YnputConnector'
 
-export const Landing = ({ nextStep, setIsConnectionLoading }) => {
+export const LandingStep = ({ nextStep, setIsConnectionLoading }) => {
   const [markdown, setMarkdown] = useState('')
 
   useEffect(() => {
