+++ src/pages/OnBoarding/Step/index.jsx
@@ -1,2 +1,2 @@
-export { Landing } from './Landing'
-export { Package } from './Package'
+export { LandingStep } from './LandingStep'
+export { PresetStep } from './PresetStep'
